feat: Add option to clean falsy values in object clean
Add a new option to `objectClean` function which allows to clean falsy values (including `undefined` and `null`) from an object when set to `true`. This is achieved by deleting all the falsy values from the object when the option is enabled.

The change also includes updates to the corresponding tests.

Signed-off-by: Avior <f.bouillon@aptatio.com>